finished adjusting tests
