+++ pymove/__init__.py
@@ -9,7 +9,9 @@ spatial-temporal data
 
 from ._version import __version__
 from .core import grid
-from .core.dataframe import DaskMoveDataFrame, MoveDataFrame, PandasMoveDataFrame
+from .core.dask import DaskMoveDataFrame
+from .core.dataframe import MoveDataFrame
+from .core.pandas import PandasMoveDataFrame
 from .models.pattern_mining import clustering
 from .preprocessing import compression, filters, segmentation, stay_point_detection
 from .semantic import semantic
