+++ setup.cfg
@@ -11,7 +11,7 @@ exclude =
 multi_line_output = 3
 include_trailing_comma = True
 line_length = 90
-known_third_party = IPython,branca,dask,folium,holidays,ipywidgets,joblib,matplotlib,numpy,pandas,psutil,psycopg2,pymongo,shapely,sklearn,tqdm
+known_third_party = IPython,branca,dask,dateutil,folium,holidays,ipywidgets,joblib,matplotlib,numpy,pandas,psutil,psycopg2,pymongo,shapely,sklearn,tqdm
 
 [aliases]
 test = pytest
